% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contourfilled_data.R
\name{contourfilled.data}
\alias{contourfilled.data}
\title{Makes filled contour plot from data without sidebar by interpolating with Gaussian process, uses contourfilled}
\usage{
contourfilled.data(x, y = NULL, z = NULL, xcontlim = NULL,
  ycontlim = NULL, ...)
}
\arguments{
\item{x}{either just x data, x and y data, or x, y and z data}

\item{y}{either y data, z data, or null}

\item{z}{either z data or null}

\item{xcontlim}{x limits for the contour plot}

\item{ycontlim}{y limits for the contour plot}

\item{...}{passed to contourfilled.func}
}
\description{
Makes filled contour plot from data without sidebar by interpolating with Gaussian process, uses contourfilled
}
\examples{
x <- runif(20)
y <- runif(20)
z <- exp(-(x-.5)^2-5*(y-.5)^2)
contourfilled.data(x,y,z)
}
\references{
[1] filled.contour R function, copied function but removed part for sidebar

[2] http://stackoverflow.com/questions/16774928/removing-part-of-a-graphic-in-r, answer by P Lapointe
}

