% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contourfilled.R
\name{contourfilled.func}
\alias{contourfilled.func}
\title{Makes filled contour plot from function without sidebar, uses contourfilled}
\usage{
contourfilled.func(fn0, n = 100, xcontlim = c(0, 1), ycontlim = c(0, 1),
  mainminmax = T, batchmax = 1, out.col.name = NULL, ...)
}
\arguments{
\item{fn0}{function to plot, first argument must be two-dimensional}

\item{n}{number of points in each dimension}

\item{xcontlim}{x limits for the contour plot}

\item{ycontlim}{y limits for the contour plot}

\item{mainminmax}{whether the min and max values should be shown in the title of plot}

\item{batchmax}{number of datapoints that can be computed at a time}

\item{out.col.name}{if a column needs to be selected from the function, specify it}

\item{...}{Passed to contourfilled}
}
\description{
Makes filled contour plot from function without sidebar, uses contourfilled
}
\examples{
contourfilled.func(function(x){x[1]*x[2]})
contourfilled.func(function(x)(exp(-(x[1]-.5)^2-5*(x[2]-.5)^2)))
}
\references{
[1] filled.contour R function, copied function but removed part for sidebar

[2] http://stackoverflow.com/questions/16774928/removing-part-of-a-graphic-in-r, answer by P Lapointe
}

