% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contourfilled_multicolor_title.R
\name{multicolor.title}
\alias{multicolor.title}
\title{Makes plot title using specified colors for the text}
\usage{
multicolor.title(main, col.main, collapse = "")
}
\arguments{
\item{main}{Text to put in main title of plot}

\item{col.main}{Colors to use for the text}

\item{collapse}{What to put between elements of main, defaults to "" but " " might be appropriate}
}
\description{
Makes plot title using specified colors for the text
}
\examples{
plot(1:4)
multicolor.title(c('Black, ','red, ','green'),c(1,2,3))
}

